+++ src/index.js
@@ -4,7 +4,7 @@ const express = require('express');
 const mongoose = require('mongoose');
 const authRoutes = require('./routes/authRoutes');
 const requireAuth = require('./middlewares/requireAuth');
-const db = require('./mysql.connector');
+const db = require('./database/mysqlConnector');
 const app = express();
 app.use(express.json());
 app.use(express.urlencoded({extended: true}));
